Renamed search_history controller to searchHistory
