+++ seasoned_api/src/webserver/controllers/user/searchHistory.js
